==> humpback_risk/3_Entanglement_report_mnpred.Rmd <==
---
title: "Compare humpback predictions with entanglement reports"
author: "Sam Woodman"
date: "6/22/2020"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Introduction

This document examines the corresponding humpback predictions for each entanglement report. This includes 1) the value of the corresponding humpback prediction and 2) a map with a point representing the entanglement report. 

```{r intro, message=FALSE}
library(classInt)
# library(lubridate)
library(maps)
library(RColorBrewer)
library(readxl)
library(tidyverse)
library(sf)

source(here::here("humpback_risk/plot_raimbow.R"))
file.ent <- "C:/SMW/RAIMBOW/raimbow-local/Data/Entanglements/Entanglements_known_gear_set_location.xlsx"
file.grid <- "C:/SMW/RAIMBOW/raimbow-local/Data/5x5 km grid shapefile/five_km_grid_polys_geo.shp"

# flag1 = 1: Only used CA-DC, know-set records (Sheet 1). 
# flag1 = 2: Use all humpack entanglement records, regardless of fishery (Sheet 2)
flag1 <- 1

# Save plots to file
flag.save <- FALSE
```

First we load and process the entanglement data and humpback predictions. Because we are only working with entanglement records, we load all humpback entanglement records and create a column indicating if this was a CA Dungeness crab fishery entanglement with a known gear set location. 

Also of note - currently we are using the Mn predictions overlaid on Blake's 5km grid, and only using entanglement records/predictions after November 2009. It may be good to do this analysis using the 3km predictions all the way back to 2005. 

```{r load}
# Load Mn preds and convert to long form
load("C:/SMW/RAIMBOW/raimbow-local/RDATA_files/Whale_risk.Rdata")
rm(df.key) #, fish.all.num, risk.all.dens)

z <- humpback.abund %>% #humpback.all.dens %>% 
  gather(key = "col_name", value = "H_Abund", 
         -GRID5KM_ID, -LONGITUDE, -LATITUDE, -area_km_lno) %>% 
  mutate(Year = as.numeric(substr(col_name, 4, 7)), 
         Month = as.numeric(substr(col_name, 9, 10)), 
         Year_Month = paste(Year, sprintf("%02d", Month), sep = "_")) %>% 
  select(GRID5KM_ID, #LONGITUDE, LATITUDE, area_km_lno, 
         Year_Month, Year, Month, H_Abund)

# z2 <- read_csv("../raimbow-local/Data/Humpback 3km models/Model1_PredictionGrids/WEAR3km_76_2005-01-01to2019-08-14_daily_dens.csv")


x.orig1 <- read_xlsx(file.ent, sheet = 1)
x.orig2 <- read_xlsx(file.ent, sheet = 2) %>% mutate(CA_DC_kgsl = CaseID %in% x.orig1$CaseID)

stopifnot(
  sum(x.orig2$CA_DC_kgsl) == nrow(x.orig1), 
  all(x.orig1$CaseID %in% x.orig2$CaseID)
)

# 'CA_DC_kgsl' means CA-DC entanglement with known gear set location
x <- x.orig2 %>% 
  filter(Common_Name %in% c("Humpback", "Humpback Whale"),
         Year >= 2010 | (Year == 2009 & Month >= 11)) %>%
  select(CaseID, Year, Month, Day, Report_Lat, Report_Long, CA_DC_kgsl)
```

Load 5km grid, make `sf` objects of the Mn preds and entanglement data, and visualize entanglement report locations

```{r, message=FALSE}
grid.5km <- st_read(file.grid) %>% select(GRID5KM_ID)

humpback.sf <- humpback.abund %>% 
  left_join(grid.5km, by = "GRID5KM_ID") %>% 
  select(starts_with("Mn_"), geometry) %>% 
  st_as_sf(agr = "constant")

x.sf <- st_as_sf(x, coords = c("Report_Long", "Report_Lat"), crs = 4326)

plot(st_geometry(humpback.sf), axes = TRUE, main = "Confirmed CA-DC entanglement reports")
plot(st_geometry(filter(x.sf, !CA_DC_kgsl)), add = TRUE, col = "orange")
plot(st_geometry(filter(x.sf, CA_DC_kgsl)), add = TRUE, col = "red")
legend("topright", title = "Mn entanglement", legend = c("All fisheries", "CA-DC known-gear-set-location"), 
       pch = 1, col = c("orange", "red"))
```

Determine grid cell ID of each entanglement

```{r}
x.sf.int <- x.sf %>% 
  st_intersection(grid.5km) %>% 
  left_join(z, by = c("Year", "Month", "GRID5KM_ID")) %>% 
  filter(!is.na(H_Abund)) %>%
  arrange(Year, Month)

summary(z$H_Abund)
summary(x.sf.int$H_Abund)
```

## Plot maps of months with entanglement reports

Prep work copied from 'Whale_risk_maps.Rmd'. Then plot!

```{r map, fig.width=4, fig.height=7, eval=flag.save}
# Code only run if flag.save is TRUE

### Prep
# Get/set break points and color palettes
n.breakpts <- 5
col.pal <- rev(brewer.pal(n.breakpts, "YlGnBu"))
h.vals.all <- unlist(st_drop_geometry(humpback.sf))

set.seed(42)
h.br <- classIntervals(na.omit(h.vals.all), n.breakpts, style = "fisher", samp_prop = 0.01)$brks
h.br[1] <- 0
h.br[n.breakpts + 1] <- max(h.vals.all, na.rm = TRUE)

# Get other objects as needed
map.contours <- st_read("C:/SMW/RAIMBOW/raimbow-local/Data/West_coast_bathy/West_Coast_geo.shp")
map.contours.100m <- map.contours %>% filter(Contour == -100) %>% st_geometry()
map.contours.200m <- map.contours %>% filter(Contour == -200) %>% st_geometry()

map.base <- st_geometry(st_as_sf(maps::map('world', plot = FALSE, fill = TRUE)))


if (flag1 == 1) {
  path.plots <- "C:/SMW/RAIMBOW/raimbow-local/Plots/Entanglementreport_Mnpreds/Entanglement_CADC_Mnpreds/"
  x.sf.int <- x.sf.int %>% filter(CA_DC_kgsl)
} else if (flag1 == 2) {
  path.plots <- "C:/SMW/RAIMBOW/raimbow-local/Plots/Entanglementreport_Mnpreds/Entanglement_allMn_Mnpreds/"
}


### Make maps!
for(i in sort(unique(x.sf.int$Year_Month))) {
  x.curr <- x.sf.int %>% filter(Year_Month == i)
  
  h.sf <- humpback.sf %>% select(contains(i))
  if (ncol(h.sf) != 2) {
    warning("No Mn predictions for year-month: ", i)
    
  } else {
    png(paste0(path.plots, "Ent_Mn_", i, ".png"), height = 7, width = 4, 
        units = "in", res = 300)
    
    plot_raimbow(
      h.sf, 1, NULL, map.base, 
      map.b1 = map.contours.100m, map.b2 = map.contours.200m, 
      col.pal = col.pal, col.breaks = h.br, 
      asp = 0, ylim = c(32, 48), xaxt = "n", 
      main = paste("Humpback abundance", i)
    )
    legend_raimbow(h.br, "%0.2f", fill = rev(col.pal), cex = 1.2, title = "Whales")
    legend.raimbow.bathy()
    
    plot(st_geometry(x.curr), add = TRUE, col = "red", cex = 1.5, lwd = 2)
    legend(x = par("usr")[2], y = 43, legend = paste0("Entanglement\nreport, n=", nrow(x.curr)), 
           xjust = 1, pch = 1, col = "red", pt.cex = 1.5)
    
    graphics::box()
    
    dev.off()
  }
}
```

## Compare Mn entanglement record with prediction value

```{r threshold, fig.width=9}
x.int <- x.sf.int %>% 
  st_drop_geometry() %>% 
  mutate(Year_Month = factor(Year_Month), 
         In_Season = factor(ifelse(between(Month, 3, 7), 19, 15), 
                            labels = c("August-February", "March-July")))

x.int %>% 
  ggplot(aes(Year_Month, H_Abund, colour = CA_DC_kgsl, shape = In_Season)) +
  geom_point(size = 3) + 
  # scale_y_continuous(limits = c(0, 0.08)) +
  ggtitle("Humpback prediction value for each humpback entanglement report") + 
  ylab("Humpback abundance (whales)") + 
  guides(color = guide_legend(title = "CA-DC known gear set location"), 
         shape = guide_legend(title = "Report month")) + 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.4))
```
