==> humpback_risk/4_Entanglements_risk.Rmd <==
---
title: "Relating entanglement reports to risk"
author: "Sam Woodman"
date: '`r format(Sys.Date(), "%d %B %Y")`'
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
if (Sys.info()["nodename"] == "SWC-SWOODMAN-L") knitr::opts_knit$set(root.dir = "../")
```

## Introduction

Goal: Examine monthly risk histograms relative to risk for entanglement reports, using lookback window to account for the uncertainty in when the actual entanglement occurred.

```{r, message=FALSE}
library(gridExtra)
library(lubridate)
library(readxl)
library(tidyverse)

source("User_script_local.R")
source("humpback_risk/funcs_whale_risk_timeseries.R")

if (user == "JS") {
  file.ent
  
} else if (user == "SMW") {
  flag.save <- FALSE
  file.ent <- "../raimbow-local/Data/Entanglements/Entanglements_known_gear_set_location.xlsx"
  path.rdata <- "../raimbow-local/RDATA_files/"
  path.plots1 <- "../raimbow-local/Plots/Entanglements_risk/Entanglements_timeseries/"
  path.plots2 <- "../raimbow-local/Plots/Entanglements_risk/Entanglements_county_timeseries/"
  
} else {
  stop("User not recognized")
}

# Options: "ca-" for CA- regions, or "county" for county-level
reg.flag <- "ca-"
hist.width <- 0.05
```

## Data import and processing

Read in entanglement data, and select relevant columns. Then if necessary, assign each entanglement to a region, using Gear_Set_County to determine region.

```{r}
### Read in and process entanglement data
x.orig <- read_xlsx(file.ent)

x <- x.orig %>% 
  select(CaseID, Year, Month, Report_Lat, Report_Long, Report_County, 
         Gear_Set_County, Gear_Set_Location_State, Timing)


# Get year-month factor levels
ym.levels <- tibble(
  year = c(rep(2009, 2), sapply(2010:2017, rep, 12), rep(2018, 7)),
  month = sprintf("%02d", c(11, 12, rep(1:12, 8), 1:7))
) %>% 
  mutate(ym = paste(year, month, sep = "-"))


# Set variables depending on region level
stopifnot(reg.flag %in% c("ca-", "county"))
if (reg.flag == "ca-") {
  # Large regions
  load(paste0(path.rdata, "Whale_risk_timeseries.Rdata"))
  path.plots <- path.plots1
  fig.h <- 8
  y.max.ent <- 15
  
  reg.list <- list(
    "Del_Norte" = "CA-N", "Humboldt" = "CA-N", "Mendocino" = "CA-N", 
    "Sonoma" = "CA-Cen", "Marin" = "CA-Cen", "San_Francisco" = "CA-Cen",
    "San_Mateo" = "CA-Cen", "Santa_Cruz" = "CA-Cen", "Monterey" = "CA-Cen", 
    "San_Luis_Obispo" = "CA-SCen", "Santa_Barbara" = "CA-SCen", 
    "Ventura" = "CA-S", "Los_Angeles" = "CA-S", "Orange" = "CA-S", 
    "San_Diego" = "CA-S"
  ) # Monterey definitely has some area in CA-SCen too
  
  county.reg <- data.frame(region = t(as.data.frame(reg.list)), 
                           stringsAsFactors = FALSE) %>% 
    rownames_to_column(var = "county") %>% 
    mutate(county = gsub("_", " ", county, ))
  
  reg.levels <- c("WA", "OR", "CA-N", "CA-Cen", "CA-SCen")
  x <- left_join(x, county.reg, by = c("Gear_Set_County" = "county"))
  
  
} else if (reg.flag == "county") {
  # County-level regions
  load(paste0(path.rdata, "Whale_risk_county_timeseries.Rdata"))
  path.plots <- path.plots2
  fig.h <- 10
  y.max.ent <- 10
  
  county.no <- c("Ventura", "Los_Angeles", "Orange", "San_Diego")
  reg.levels <- levels(all.df.summ$region)[!(levels(all.df.summ$region) %in% county.no)]
  x <- x %>% mutate(region = gsub(" ", "_", Gear_Set_County))
  
  
} else {
  stop("reg.flag must be eone of \"ca-\" and \"county\"")
}

# Finish entanglement record prep
x <- x %>% 
  mutate(Gear_Set_County = as.factor(Gear_Set_County), 
         region = factor(region, levels = reg.levels), 
         ym = factor(paste(Year, sprintf("%02d", Month), sep = "-"), levels = ym.levels$ym), 
         ym_date = parse_date_time(ym, "ym")) %>% 
  filter(!is.na(ym))

reg.plot <- sort(unique(as.character(x$region)))
```

Make time series of the entanglement reports

```{r, fig.height=4, fig.width=10}
x.max <- length(levels(x$ym))
x.lab.idx <- seq(1, to = x.max, by = 3)
x.lab <- sort(levels(x$ym))[x.lab.idx]
vert.lines <- seq(0.5, to = x.max, by = 12)

gg0 <- ggplot(x, aes(x = ym, fill = region)) + 
  geom_bar() + 
  geom_vline(xintercept = vert.lines, col = "black", lwd = 0.35)

if (reg.flag == "ca-") {
  gg0 <- gg0 + scale_fill_brewer(palette = "Set1", name = "Region", drop = FALSE)
} else if (reg.flag == "county") {
  gg0 <- gg0 + scale_fill_discrete(name = "region") #, drop = FALSE)
}

gg0 + ggtitle("CA DC confirmed and known-set-location entanglement reports") +
  xlab("Year-Month") +
  ylab("Entanglements") +
  scale_x_discrete(drop = FALSE, breaks = x.lab) +
  theme(plot.title = element_text(hjust = 0.5),
        axis.text.x = element_text(angle = 90, vjust = 0.4),
        legend.justification = "top")

if (flag.save) ggsave(paste0(path.plots, "Ent_report_timeseries.png"))
```

## Compare entanglement reports with risk time series values

Now we load the time series data, and plot histograms of the risk time series values (densities) by region.

```{r, fig.width=14, fig.height=fig.h}
### Add entanglement data to time series data
# Time series data loaded above
all.df.summ.join <- all.df.summ %>% 
  mutate(region = factor(region, levels = levels(x$region))) %>% 
  select(ym, region, risk_sum_dens)

# if (flag.save) save(x, all.df.summ.join, file = paste0(path.rdata, "Ent_risk.RDATA"))


# Plot histogram of time series risk values
facet.nrow <- case_when(reg.flag == "ca-" ~ 1, reg.flag == "county" ~ 3)
ggplot(all.df.summ, aes(x = risk_sum_dens)) +
  geom_histogram(binwidth = hist.width, boundary = 0) + 
  facet_wrap(facets = vars(region), nrow = facet.nrow) +
  xlab("Risk (whales*VMS pings / km2)") + 
  ggtitle("Risk time series")

if (flag.save) ggsave(paste0(path.plots, "Risk_hist_all.png"))
```

Next we determine the corresponding risk value for each entanglement report, treating year-months in which no entanglements were reported as zero entanglement reports. Then we can make the same plots, but color-coded to indicate if there was an entanglement report during that year-month.

```{r, fig.width=14, fig.height=fig.h}
x.join <- x %>% 
  select(ym, region, CaseID) %>% 
  left_join(all.df.summ.join, by = c("ym", "region"))

# Contains all risk values and all risk-entanglement report pairs, separately
y <- all.df.summ.join %>% 
  mutate(CaseID = NA) %>% 
  bind_rows(x.join) %>% 
  mutate(e_logical = factor(ifelse(is.na(CaseID), "All", "Entanglements")), 
         ym = factor(ym), 
         region = factor(region, levels(x$region))) %>% 
  arrange(region, ym)

# Plot histogram of time series risk values, color-coded by whether or not there was an entanglement
y %>% 
  filter(region %in% reg.plot) %>%
  mutate(y_max = ifelse(e_logical == "All", 90, y.max.ent)) %>% 
  ggplot(aes(x = risk_sum_dens)) + #, fill = e_logical
  geom_histogram(binwidth = hist.width, position = "dodge", boundary = 0) +
  facet_grid(cols = vars(region), rows = vars(e_logical), 
             scales = "free_y", space = "free_y") +
  geom_blank(aes(y = y_max)) + 
  # facet_wrap(facets = vars(region), nrow = facet.nrow) +
  scale_y_continuous(breaks = seq(0, 100, by = 5)) +
  guides(fill = guide_legend(title = "Risk values")) + 
  xlab("Risk (whales*VMS pings / km2)") + 
  ggtitle("Regional (monthly) risk - all risk values vs risk values for entanglement reports - original (actual) report timing")

if (flag.save) ggsave(paste0(path.plots, "Risk_hist_orig.png"))
```

## Compare entanglement reports with risk time series values - pt2

Now, we wish to take into account the fact that we do not know exactly when the entanglement occurred. To do this, we use a 'lookback window', meaning we look back at the maximum risk value that occurred in that region within the past 1 month, 2 months, etc. 

TODO: we also want to look at the sum of the risk values that occurred in that region within the lookback window, but not sure yet how to make a meaningful plot.

```{r, fig.width=14, fig.height=fig.h}
# Set size of lookback window
# lookback.mons <- 3
facet.nrow2 <- case_when(reg.flag == "ca-" ~ 1, reg.flag == "county" ~ 2)

# Loop through possible lookback windows
for (lookback.mons in 0:12) {
  # Prep entanglement records - add idx column
  x.pre <- x %>% 
    select(region, ym_date) %>% 
    mutate(ent_idx = seq_along(ym_date))
  
  # Get all ym values in lookback window, determine which ym value has max 
  x2 <- lapply(0:lookback.mons, function(i) {
    x.pre %>% 
      mutate(ym_date_curr = ym_date %m-% months(i), 
             ym_curr = substr(as.character(ym_date_curr), 1, 7))
  }) %>% 
    bind_rows() %>% 
    select(region, ym = ym_curr, ent_idx) %>% 
    left_join(all.df.summ.join, by = c("ym", "region")) %>% 
    group_by(ent_idx) %>% 
    summarise(r_max = max(risk_sum_dens, na.rm = TRUE), 
              r_sum = sum(risk_sum_dens, na.rm = TRUE), 
              r_min = min(risk_sum_dens, na.rm = TRUE), 
              ym_max = ym[which.max(risk_sum_dens)]) %>% 
    ungroup() %>% 
    right_join(x.pre, by = "ent_idx") %>% 
    select(ent_idx, ym = ym_max, region, r_max, r_sum, r_min)
  
  # Summarise entanglement count by max ym
  x2.join <- x2 %>% 
    select(ym, region, ent_idx) %>% 
    left_join(all.df.summ.join, by = c("ym", "region"))
  
  y2 <- all.df.summ.join %>% 
    mutate(ent_idx = NA) %>% 
    bind_rows(x2.join) %>% 
    mutate(e_logical = factor(ifelse(is.na(ent_idx), "All", "Entanglements")), 
           ym = factor(ym), 
           region = factor(region, levels(x$region))) %>% 
    arrange(region, ym)
  
  # Plot
  gg2 <- y2 %>% 
    filter(region %in% reg.plot) %>%
    mutate(y_max = ifelse(e_logical == "All", 90, 15)) %>% 
    ggplot(aes(x = risk_sum_dens)) + #, fill = e_logical)) + 
    geom_histogram(binwidth = hist.width, position = "dodge", boundary = 0) +
    facet_grid(cols = vars(region), rows = vars(e_logical), 
               scales = "free_y", space = "free_y") +
    # facet_wrap(facets = vars(region), nrow = facet.nrow2) +
    geom_blank(aes(y = y_max)) + 
    scale_fill_discrete(drop = FALSE, name = "Risk values") + 
    scale_y_continuous(breaks = seq(0, 100, by = 5)) +
    xlab("Risk (whales*VMS pings / km2)") + 
    ggtitle(paste("Regional (monthly) risk - all (n = 525) vs entanglement reports (n = 22) -", 
                  lookback.mons, 
                  "month lookback window, max risk within window"))
  
  
  # Save plot
  file.plot <- paste0(path.plots, "Risk_hist_lookback_", lookback.mons, "mons.png")
  if (flag.save) ggsave(file.plot, gg2)
  
  # print(gg2)
  rm(x.pre, x2, x2.join, y2, gg2)
}
```

Experimental: for each (some?) lookback window, how far back is the max risk?

```{r}
x.out <- x.orig %>% 
  left_join(select(x, CaseID, region), by = "CaseID")

for(lookback.mons in seq(0, 12, by = 3)) {
  # Prep entanglement records - add ent_idx column
  x.pre <- x %>% 
    mutate(ent_idx = seq_along(ym_date))
  
  # Get all ym values in lookback window, determine which ym value has max 
  x2 <- lapply(0:lookback.mons, function(i) {
    x.pre %>% 
      mutate(ym_date_curr = ym_date %m-% months(i), 
             ym_curr = substr(as.character(ym_date_curr), 1, 7))
  }) %>% 
    bind_rows() %>% 
    select(region, ym = ym_curr, ent_idx) %>% 
    left_join(all.df.summ.join, by = c("ym", "region")) %>% 
    group_by(ent_idx) %>% 
    summarise(r_max = max(risk_sum_dens, na.rm = TRUE), 
              r_sum = sum(risk_sum_dens, na.rm = TRUE), 
              ym_max = ym[which.max(risk_sum_dens)]) %>% 
    ungroup() %>% 
    right_join(x.pre, by = "ent_idx") %>% 
    select(CaseID, ym_max) %>% 
    set_names("CaseID", paste0("ym_max", lookback.mons))
  
  x.out <- left_join(x.out, x2, by = "CaseID")
  rm(x.pre, x2)
}

if (flag.save) write.csv(x.out, file = paste0(path.plots, "Ent_ymmax.csv"), row.names = FALSE)

table(x.out$ym_max0, as.character(x.out$region))
table(x.out$ym_max3, as.character(x.out$region))
table(x.out$ym_max6, as.character(x.out$region))
table(x.out$ym_max9, as.character(x.out$region))
table(x.out$ym_max12, as.character(x.out$region))
```

## Cumulative risk within lookback window

Calculate sum of risk values in lookback window instead of max of risk values in lookback window. This also involves summing the risk values in the lookback window for all of the time series values.

```{r, fig.width=14, fig.height=fig.h}
# Set size of lookback window
# lookback.mons <- 3
#set above: # facet.nrow2 <- case_when(reg.flag == "ca-" ~ 1, reg.flag == "county" ~ 2) 
z <- y

# Loop through possible lookback windows
for (lookback.mons in 0:12) {
  
  # Prep stuff - add idx column
  z.pre <- z %>% 
    mutate(ym_date = parse_date_time(ym, "ym"), 
           ent_idx = seq_along(ym_date)) %>% 
    select(region, ym, ym_date, ent_idx, e_logical)
  
  # Get all ym values in lookback window, sum values
  z2 <- lapply(0:lookback.mons, function(i) {
    z.pre %>% 
      select(-ym, -e_logical) %>% 
      mutate(ym_date_curr = ym_date %m-% months(i), 
             ym_curr = substr(as.character(ym_date_curr), 1, 7))
  }) %>% 
    bind_rows() %>% 
    select(region, ym = ym_curr, ent_idx) %>% 
    left_join(all.df.summ.join, by = c("ym", "region")) %>% 
    group_by(ent_idx) %>% 
    summarise(r_sum = sum(risk_sum_dens, na.rm = TRUE), 
              ym_max = ym[which.max(risk_sum_dens)]) %>% 
    ungroup() %>% 
    right_join(z.pre, by = "ent_idx") %>% 
    select(ent_idx, ym, region, risk_sum_dens = r_sum, e_logical)
  
  # Plot
  gg2 <- z2 %>% 
    filter(region %in% reg.plot) %>%
    mutate(y_max = ifelse(e_logical == "All", 90, 15)) %>% 
    ggplot(aes(x = risk_sum_dens)) + #, fill = e_logical)) + 
    geom_histogram(binwidth = hist.width, position = "dodge", boundary = 0) +
    facet_grid(cols = vars(region), rows = vars(e_logical), 
               scales = "free_y", space = "free_y") +
    geom_blank(aes(y = y_max)) + 
    scale_fill_discrete(drop = FALSE, name = "Risk values") + 
    scale_y_continuous(breaks = seq(0, 100, by = 5)) +
    xlab("Risk (whales*VMS pings / km2)") + 
    ggtitle(paste("Regional (monthly) risk - all (n = 525) vs entanglement reports (n = 22) -", 
                  lookback.mons, 
                  "month lookback window, summed risk within window"))
  
  
  # Save plot
  file.plot <- paste0(path.plots, "Risk_hist_lookbacksum_", lookback.mons, "mons.png")
  if (flag.save) ggsave(file.plot, gg2)
  
  # print(gg2)
}
```
