==> humpback_risk/4_Whale_risk_county_management.Rmd <==
---
title: "Humpback risk - management scenarios - county"
author: "Sam Woodman"
date: '`r format(Sys.Date(), "%d %B %Y")`'
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
if (Sys.info()["nodename"] == "SWC-SWOODMAN-L") knitr::opts_knit$set(root.dir = "../")
```

## Introduction

The goal of this document is to examine the change in risk under different management scenarios, e.g. early or late season closures.

```{r prep, message = FALSE}
library(tidyverse)

load("../raimbow-local/RDATA_files/Whale_risk_county_timeseries.Rdata")
# source("Whale_risk_mgt_funcs.R")

flag.save <- FALSE
path.plots <- "../raimbow-local/Plots/Management/"
```

To do this, we first must summarize the original risk (i.e., with no management scenarios) by region and fishing season

```{r orig}
county.rm <- c("Ventura", "Los_Angeles", "Orange", "San_Diego")
county.levels <- levels(all.df.summ$region)[!(levels(all.df.summ$region) %in% county.rm)]

x <- all.df.summ %>% 
  mutate(region = factor(region, levels = county.levels)) %>% 
  arrange(yr, mon, region) %>% 
  mutate(fish_season = cumsum(mon == 11 & region == "WA")) %>% 
  select(region, ym, yr, mon, fish_season, everything()) %>% 
  arrange(region, yr, mon)

x.orig <- x %>% 
  group_by(region, fish_season) %>% 
  summarise(yr = min(yr), 
            risk_total_season = sum(risk_sum_total), 
            risk_dens_season = sum(risk_sum_dens)) %>% 
  ungroup()

head(x.orig)
```

## Make function(s)

Functions used throughout this document

```{r funcs}
# TODO: Use "funcs..." instead



# Sum risk after filtering for desired months
mng_risk_sum <- function(x1, x2, mon.keep) {
  x1 %>%
    filter(mon %in% mon.keep) %>%
    group_by(region, fish_season) %>%
    summarise(risk_total_season = sum(risk_sum_total),
              risk_dens_season = sum(risk_sum_dens)) %>%
    ungroup() %>%
    mutate(risk_total_ratio = risk_total_season / x2$risk_total_season,
           risk_dens_ratio = risk_dens_season / x2$risk_dens_season,
           risk_total_diff = risk_total_season - x2$risk_total_season,
           risk_dens_diff = risk_dens_season - x2$risk_dens_season)
}

# Plot of percentage of risk remaining for the season after applying scenario
mng_plot_perc <- function(x1, x2, x.title, x.xlab, x.ylab) {
  ggplot(x1, aes(fish_season, risk_dens_ratio, colour = region, group = region)) +
    geom_point() +
    geom_path() +
    # scale_colour_brewer(palette = "Set1", name = "Region") +
    ggtitle(x.title) +
    xlab(x.xlab) +
    ylab(x.ylab) +
    coord_cartesian(ylim = c(0, 1)) +
    scale_x_continuous(breaks = seq_along(unique(x2$fish_season)),
                       labels = unique(x2$yr))
}


# Plot of summed risk density for the year with and without management scenario
mng_plot_bar <- function(x1, x2, v1, x.title, x.ylab, f.nrow = 2) {
  v1 <- enquo(v1)

  x.new <- bind_rows(
    x2 %>% select(region, fish_season, risk_total_season, risk_dens_season),
    x1 %>% select(region, fish_season, risk_total_season, risk_dens_season)
  ) %>%
    mutate(orig = factor(c(rep("Original", nrow(x.orig)), rep("Closure", nrow(x.late)))))


  ggplot(x.new, aes(fish_season, !!v1, fill = orig)) +
    geom_col(position = "dodge") +
    facet_wrap(facets = vars(region), nrow = f.nrow) +
    guides(fill = guide_legend(title = "Scenario")) +
    ggtitle(x.title) +
    xlab("Fishing season start year") +
    ylab(x.ylab) +
    scale_x_continuous(breaks = seq_along(unique(x2$fish_season)),
                       labels = unique(x2$yr)) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.4))
}
```

## Late season closure, e.g. CA-CBD settlement

This scenario is meant to test how risk would have changed if the CA-CBD settlement, closing the DC fishery April 15, had been in place beginning in 2009. Because we are working with monthly-summarized data, we will consider the fishery to be open through April.

```{r late}
x.late <- mng_risk_sum(x, x.orig, c(11, 12, 1:4))

mng_plot_perc(
  x.late, x.orig, 
  x.title = "Percentage of risk remaining with closing DC fishery May 1", 
  x.xlab = "Fishing season start year", 
  x.ylab = "Percentage"
)

if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_close05.png"))


### Bar plot(s)
mng_plot_bar(
  x.late, x.orig, risk_dens_season, 
  "Summed risk (density) - closing DC fishery May 1", 
  "Risk (whales*VMS pings / km2)", f.nrow = 3
)
if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_close05_bar.png"), 
                      width = 10, height = 8)

mng_plot_bar(
  x.late, x.orig, risk_total_season, 
  "Summed risk (total) - closing DC fishery May 1", 
  "Risk (whales*VMS pings)", f.nrow = 3
)
if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_close05_total.png"), 
                      width = 10, height = 8)
```

The only fishing season in which this closure had a strong impact on regions in which most confirmed entanglements have happened is 2015-16, when the fishery was closed unitl March and thus 

## Early season closure, e.g. 2019-20 fishing season

Opening fishery Dec 1

```{r early1}
x.early1 <- mng_risk_sum(x, x.orig, c(12, 1:10))

mng_plot_perc(
  x.early1, x.orig, 
  x.title = "Percentage of risk remaining with opening DC fishery Dec 1", 
  x.xlab = "Fishing season start year", x.ylab = "Percentage"
)

if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_open12.png"))


### Bar plot(s)
mng_plot_bar(
  x.early1, x.orig, risk_dens_season, 
  "Summed risk (density) - opening DC fishery Dec 1", 
  "Risk (whales*VMS pings / km2)", f.nrow = 3
)
if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_open12_bar.png"), 
                      width = 10, height = 8)

mng_plot_bar(
  x.early1, x.orig, risk_total_season, 
  "Summed risk (total) - opening DC fishery Dec 1", 
  "Risk (whales*VMS pings)", f.nrow = 3
)
if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_open12_bar_total.png"), 
                      width = 10, height = 8)
```

Opening fishery Jan 1

```{r early2}
x.early2 <- mng_risk_sum(x, x.orig, c(1:10))

mng_plot_perc(
  x.early2, x.orig, 
  x.title = "Percentage of risk remaining with opening DC fishery Jan 1", 
  x.xlab = "Fishing season start year", x.ylab = "Percentage"
)

if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_open01.png"))


### Bar plot(s)
mng_plot_bar(
  x.early2, x.orig, risk_dens_season, 
  "Summed risk (density) - opening DC fishery Jan 1", 
  "Risk (whales*VMS pings / km2)", f.nrow = 3
)
if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_open01_bar.png"), 
                      width = 10, height = 8)

mng_plot_bar(
  x.early2, x.orig, risk_total_season, 
  "Summed risk (total) - opening DC fishery Jan 1", 
  "Risk (whales*VMS pings)", f.nrow = 3
)
if (flag.save) ggsave(filename = paste0(path.plots, "county_Mn_DC_open01_bar_total.png"), 
                      width = 10, height = 8)
```
