==> humpback_risk/4_Whale_risk_management.Rmd <==
---
title: "Humpback risk - management scenarios"
author: "Sam Woodman"
date: '`r format(Sys.Date(), "%d %B %Y")`'
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# if (Sys.info()["nodename"] == "SWC-SWOODMAN-L") knitr::opts_knit$set(root.dir = "../")
```

## Introduction

The goal of this document is to examine the change in risk under different management scenarios, e.g. early or late season closures.

```{r prep, message = FALSE}
library(scales)
library(tidyverse)

load("C:/SMW/RAIMBOW/raimbow-local/RDATA_files/Whale_risk_timeseries.Rdata")
source("funcs_whale_risk_mgmt.R")

flag.save <- TRUE
wa.inc <- TRUE
path.plots <- if (wa.inc) {
  "C:/SMW/RAIMBOW/raimbow-local/Plots/Management/"
} else {
  "C:/SMW/RAIMBOW/raimbow-local/Plots/Management_noWA/"
}
```

To do this, we first must summarize the original risk (i.e., with no management scenarios) by region and fishing season

```{r orig}
x <- all.df.summ %>% 
  arrange(yr, mon, region) %>% 
  mutate(fish_season = cumsum(mon == 11 & region == "WA")) %>% 
  select(region, ym, yr, mon, fish_season, everything()) %>% 
  arrange(region, yr, mon)

x.orig <- x %>% 
  group_by(region, fish_season) %>% 
  summarise(yr = min(yr), 
            risk_total_season = sum(risk_sum_total), 
            risk_dens_season = sum(risk_sum_dens), 
            .groups = "drop")

# TODO: update as needed, No OR data for 2018
x.orig.tmp <- x.orig %>% 
  filter(!(region %in% c("WA", "OR") & fish_season == 10))

head(x.orig, 15)
```

## Make function(s) and assign vals

The functions are now sourced above. Objects used throughout:

```{r}
xlab.all   <- "Fishing season start year (November of start year to October of next year)"
ylab.dens  <- "Risk (whales*VMS pings / km2)"
ylab.total <- "Risk (whlaes*VMS pings)"
```

## Late season closure, e.g. CA-CBD settlement

This scenario is meant to test how risk would have changed if the CA-CBD settlement, closing the DC fishery April 15, had been in place beginning in 2009. Because we are working with monthly-summarized data, we will consider the fishery to be open through April.

```{r late}
# x.close05 <- mgmt_risk_sum(x, x.orig, c(11, 12, 1:4), wa.flag = wa.inc)
# 
# # ### Line plot(s)
# # mgmt_plot_perc(
# #   x.close05, x.orig, 
# #   x.title = "Percentage of risk remaining with closing DC fishery May 1", 
# #   x.xlab = xlab.all, x.ylab = "Percentage"
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close05.png"))
# # 
# # mgmt_plot_diff(
# #   x.close05, x.orig, 
# #   x.title = "Difference in risk for fishing season with closing DC fishery May 1", 
# #   x.xlab = xlab.all, x.ylab = ylab.dens
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close05_diff.png"))
# 
# 
# ### Bar plot(s)
# # mgmt_plot_bar(
# #   x.close05, x.orig, risk_dens_season, 
# #   "Summed risk (density) - closing DC fishery May 1", ylab.dens
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close05_bar.png"))
# # 
# # mgmt_plot_bar(
# #   x.close05, x.orig, risk_total_season, 
# #   "Summed risk (total) - closing DC fishery May 1", ylab.total
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close05_bar_total.png"))
# 
# mgmt_plot_bardot(
#   x.close05, x.orig, risk_dens_season, 
#   "Summed risk (density) - closing DC fishery May 1", ylab.dens, wa.flag = wa.inc
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close05_bardot_dens.png"))
# 
# mgmt_plot_bardot(
#   x.close05, x.orig, risk_total_season, 
#   "Summed risk (total) - closing DC fishery May 1", ylab.total, wa.flag = wa.inc
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close05_bardot_total.png"))
```

Close April 1

```{r late2}
# TODO: update as needed, No OR data for 2018
x.close04 <- mgmt_risk_sum(x, x.orig, c(11, 12, 1:3), wa.flag = wa.inc) %>% 
  filter(!(region %in% c("WA", "OR") & fish_season == 10))

# ### Line plot(s)
# mgmt_plot_perc(
#   x.close04, x.orig, 
#   x.title = "Percentage of risk remaining with closing DC fishery April 1", 
#   x.xlab = xlab.all, x.ylab = "Percentage"
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close04.png"))
# 
# mgmt_plot_diff(
#   x.close04, x.orig, 
#   x.title = "Difference in risk for fishing season with closing DC fishery April 1", 
#   x.xlab = xlab.all, x.ylab = ylab.dens
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close04_diff.png"))


### Bar plot(s)
# mgmt_plot_bar(
#   x.close04, x.orig, risk_dens_season, 
#   "Summed risk (density) - closing DC fishery April 1", ylab.dens
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close04_bar.png"))
# 
# mgmt_plot_bar(
#   x.close04, x.orig, risk_total_season, 
#   "Summed risk (total) - closing DC fishery April 1", ylab.total
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close04_bar_total.png"))

mgmt_plot_bardot(
  x.close04, x.orig.tmp, risk_dens_season, 
  "Summed risk (density) - closing DC fishery April 1", ylab.dens, wa.flag = wa.inc
)
if (flag.save) ggsave(filename = "Mn_DC_close04_bardot_dens.png", path = path.plots)

mgmt_plot_bardot(
  x.close04, x.orig.tmp, risk_total_season, 
  "Summed risk (total) - closing DC fishery April 1", ylab.total, wa.flag = wa.inc
)
if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_close04_bardot_total.png"))
```

The only fishing season in which this closure had a strong impact on regions in which most confirmed entanglements have happened is 2015-16, when the fishery was closed unitl March and thus 

## Early season closure, e.g. 2019-20 fishing season

Opening fishery Dec 1

```{r early1}
# x.open12 <- mgmt_risk_sum(x, x.orig, c(12, 1:10), wa.flag = wa.inc)
# 
# # ### Line plots
# # mgmt_plot_perc(
# #   x.open12, x.orig, 
# #   x.title = "Percentage of risk remaining with opening DC fishery Dec 1", 
# #   x.xlab = xlab.all, x.ylab = "Percentage"
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open12.png"))
# # 
# # mgmt_plot_diff(
# #   x.open12, x.orig, 
# #   x.title = "Difference in risk for fishing season with opening DC fishery Dec 1", 
# #   x.xlab = xlab.all, x.ylab = ylab.dens
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open12_diff.png"))
# 
# 
# ### Bar plot(s)
# # mgmt_plot_bar(
# #   x.open12, x.orig, risk_dens_season, 
# #   "Summed risk (density) - opening DC fishery Dec 1", ylab.dens
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open12_bar.png"))
# # 
# # mgmt_plot_bar(
# #   x.open12, x.orig, risk_total_season, 
# #   "Summed risk (total) - opening DC fishery Dec 1", ylab.total
# # )
# # if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open12_bar_total.png"))
# 
# mgmt_plot_bardot(
#   x.open12, x.orig, risk_dens_season, 
#   "Summed risk (density) - opening DC fishery Dec 1", ylab.dens, wa.flag = wa.inc
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open12_bardot_dens.png"))
# 
# mgmt_plot_bardot(
#   x.open12, x.orig, risk_total_season, 
#   "Summed risk (total) - opening DC fishery Dec 1", ylab.total, wa.flag = wa.inc
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open12_bardot_total.png"))
```

Opening fishery Jan 1

```{r early2}
# TODO: update as needed, No OR data for 2018
x.open01 <- mgmt_risk_sum(x, x.orig, c(1:10), wa.flag = wa.inc) %>% 
  filter(!(region %in% c("WA", "OR") & fish_season == 10))

# ### Line plots
# mgmt_plot_perc(
#   x.open01, x.orig, 
#   x.title = "Percentage of risk remaining with opening DC fishery Jan 1", 
#   x.xlab = xlab.all, x.ylab = "Percentage"
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open01.png"))
# 
# mgmt_plot_diff(
#   x.open01, x.orig, 
#   x.title = "Difference in risk for fishing season with opening DC fishery Jan 1", 
#   x.xlab = xlab.all, x.ylab = ylab.dens
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open01_diff.png"))


### Bar plot(s)
# mgmt_plot_bar(
#   x.open01, x.orig, risk_dens_season, 
#   "Summed risk (density) - opening DC fishery Jan 1", ylab.dens
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open01_bar.png"))
# 
# mgmt_plot_bar(
#   x.open01, x.orig, risk_total_season, 
#   "Summed risk (total) - opening DC fishery Jan 1", ylab.total
# )
# if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open01_bar_total.png"))

mgmt_plot_bardot(
  x.open01, x.orig.tmp, risk_dens_season, 
  "Summed risk (density) - opening DC fishery Jan 1", ylab.dens, wa.flag = wa.inc
)
if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open01_bardot_dens.png"))

mgmt_plot_bardot(
  x.open01, x.orig.tmp, risk_total_season, 
  "Summed risk (total) - opening DC fishery Jan 1", ylab.total, wa.flag = wa.inc
)
if (flag.save) ggsave(filename = paste0(path.plots, "Mn_DC_open01_bardot_total.png"))
```

## Other

Save files for use elsewhere

```{r, eval=FALSE}
# Code not run
save(
  x.orig, x.close04, x.close05, x.open12, x.open01, 
  file = "../raimbow-local/RDATA_files/Whale_risk_management.RDATA"
)
```
