==> humpback_risk/5_Whale_risk_timeseries_presentation.Rmd <==
---
title: "Humpback risk of entanglement - baseline analysis"
author: ""
date: '`r format(Sys.Date(), "%d %B %Y")`'
output:
   powerpoint_presentation: default
   beamer_presentation: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
if (Sys.info()["nodename"] == "SWC-SWOODMAN-L") knitr::opts_knit$set(root.dir = "../")
```

```{r}
library(gridExtra)
library(tidyverse)
source("humpback_risk/funcs_whale_risk_timeseries.R")

load("../raimbow-local/RDATA_files/Whale_risk_timeseries.Rdata")
load("../raimbow-local/RDATA_files/Whale_risk_timeseries_base.Rdata")

p.max <- max(c(x.summ$risk_sum_dens, x.summ$risk_sum_dens_basew, x.summ$risk_sum_dens_basef))
```

## Humpback risk of entanglement

```{r, fig.height=10, fig.width=10}
# <!-- Risk of entanglement is calculated as product of predicted humpback abundance and number of (non-confidential) VMS pings in a 5km grid cell. -->

p1 <- raimbow_ggplot(
  all.df.summ, risk_sum_dens, plot.main = "Risk", 
  y.lab = "Whales * VMS pings / km2", wa.flag = FALSE
)
p2 <- raimbow_ggplot(
  all.df.summ, mn_sum_dens, plot.main = "Humpback whales", 
  y.lab = "Whales / km2"
)
p3 <- raimbow_ggplot(
  all.df.summ, vms_sum_dens, plot.main = "Fishing", 
  y.lab = "VMS pings / km2", wa.flag = FALSE
)

p123 <- grid.arrange(p1, p2, p3, nrow = 3)

# ggsave(filename = "Linear_humpback_risk_regional.png", p123, path = path.plots, 
#        height = 10, width = 10)
```

## Baseline analysis

- Goal: Examine the 'contribution' of the changes in whale distribution and the changes in fishing effort to entanglement risk

- Methodology: Use 'baseline' values (average, by month and grid cell, for the 2009-2010 to 2012-2013 fishing seasons) of humpback and fishing values to calculate risk to see what would have happened in a 'normal' whale or fishing season
<!-- - Terminology:  -->
<!--     - base refers to baseline, meaning these are the ‘baseline’ values -->
<!--     - basew refers to risk values calculated using ‘baseline’ whale values -->
<!--     - basef refers to risk values calculated using ‘baseline’ fishing values -->

## Baseline - humpbacks

```{r, fig.width=10, fig.height=7}
p1 <- raimbow_ggplot(
  all.df.summ, mn_sum_dens, plot.main = "Humpback whales", 
  y.lab = "Whales / km2"
)
p2 <- raimbow_ggplot(
  x.summ, mn_sum_dens_base, plot.main = "Humpback whales (baseline)", 
  y.lab = "Whales / km2"
) + 
  scale_y_continuous(limits = c(0, max(x.summ$mn_sum_dens)))

grid.arrange(p1, p2, nrow = 2)
```

## Risk with baseline humpback

```{r, fig.width=10, fig.height=7}
p1 <- raimbow_ggplot(
  all.df.summ, risk_sum_dens, plot.main = "Risk", 
  y.lab = "Whales * VMS pings / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, p.max))
p2 <- raimbow_ggplot(
  x.summ, risk_sum_dens_basew, plot.main = "Risk - baseline humpback", 
  y.lab = "Whales (baseline) * VMS pings / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, p.max))

grid.arrange(p1, p2, nrow = 2)
```

## Baseline - fishing

```{r, fig.width=10, fig.height=7}
p1 <- raimbow_ggplot(
  all.df.summ, vms_sum_dens, plot.main = "Fishing", 
  y.lab = "VMS pings / km2", wa.flag = FALSE
)
p2 <- raimbow_ggplot(
  x.summ, vms_sum_dens_base, plot.main = "Fishing (baseline)", 
  y.lab = "VMS pings / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, max(x.summ$vms_sum_dens)))

grid.arrange(p1, p2, nrow = 2)
```

## Risk with baseline fishing

```{r, fig.width=10, fig.height=7}
p1 <- raimbow_ggplot(
  all.df.summ, risk_sum_dens, plot.main = "Risk", 
  y.lab = "Whales * VMS pings / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, p.max))
p2 <- raimbow_ggplot(
  x.summ, risk_sum_dens_basef, plot.main = "Risk - baseline fishing", 
  y.lab = "Whales * VMS pings (baseline) / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, p.max))

grid.arrange(p1, p2, nrow = 2)
```

## Risk

```{r, fig.height=10, fig.width=10}
# p.max <- max(c(x.summ$risk_sum_dens, x.summ$risk_sum_dens_basew, x.summ$risk_sum_dens_basef))
p1 <- raimbow_ggplot(
  x.summ, risk_sum_dens, plot.main = "Risk", 
  y.lab = "Whales * VMS pings / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, p.max))

p2 <- raimbow_ggplot(
  x.summ, risk_sum_dens_basew, plot.main = "Risk - baseline humpback", 
  y.lab = "Whales (baseline) * VMS pings / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, p.max))

p3 <- raimbow_ggplot(
  x.summ, risk_sum_dens_basef, plot.main = "Risk - baseline fishing", 
  y.lab = "Whales * VMS pings (baseline) / km2", wa.flag = FALSE
) + 
  scale_y_continuous(limits = c(0, p.max))

p123 <- grid.arrange(p1, p2, p3, nrow = 3)
```
